Avoid use of C# dynamic keyword
Was causing issues with the WebGL build

+++ Pop Pixie/Assets/Core Scripts/Ring Pulls/RingPullsData.cs
@@ -8,7 +8,7 @@ public class RingPullsData {
   public static bool ShouldPulse = false;
 
   public static int Amount() {
-    return GameData.Current.Fetch(
+    return (int) GameData.Current.Fetch(
       "ring-pulls", 
       orSetEqualTo: 0
     );
